test(ui): Move remaining logic tests out of source files
This concludes moving of all currently existing logic tests out of source
files.
