+++ ui/src/components/resolutions/utils.ts
@@ -200,48 +200,3 @@ export function getDisplayItems(
 
   return [];
 }
-
-// Unit tests.
-
-if (import.meta.vitest) {
-  const { describe, expect, it } = import.meta.vitest;
-
-  describe('getDisplayItems', () => {
-    it('match applied and unapplied server issue resolutions by message hash', () => {
-      const unappliedResolution = {
-        message: 'Very long issue message copied from the server',
-        messageHash: 'sha1-message-hash',
-        reason: 'BUILD_TOOL_ISSUE',
-        comment: 'Updated comment',
-        source: 'SERVER',
-      };
-      const appliedResolution = {
-        ...unappliedResolution,
-        message: 'Older persisted server message representation',
-        isDeleted: false,
-      };
-
-      const item = {
-        source: 'Analyzer',
-        message: unappliedResolution.message,
-        resolutions: [appliedResolution],
-        unappliedResolutions: [unappliedResolution],
-      } as unknown as ItemWithResolutions;
-
-      expect(getDisplayItems(item, true)).toStrictEqual([
-        {
-          key: 'SERVER:sha1-message-hash:applied',
-          resolution: appliedResolution,
-          state: 'applied',
-          showActions: false,
-        },
-        {
-          key: 'SERVER:sha1-message-hash:pending-update',
-          resolution: unappliedResolution,
-          state: 'pending-update',
-          showActions: true,
-        },
-      ]);
-    });
-  });
-}
