+++ ui/src/routes/organizations/$orgId/products/$productId/repositories/$repoId/runs/$runIndex/dependencies/-components/dependency-graph-utils.ts
@@ -102,100 +102,3 @@ export const createNodeSubtreeMatcher = (
 
   return matchesNodeSubtree;
 };
-
-//
-// Unit tests for dependency graph helpers.
-//
-
-if (import.meta.vitest) {
-  const { describe, expect, it } = import.meta.vitest;
-
-  const graph: DependencyGraph = {
-    edges: [
-      { from: 0, to: 1 },
-      { from: 1, to: 2 },
-    ],
-    nodes: [
-      { fragment: 0, linkage: 'PROJECT_DYNAMIC', packageCount: 2, pkg: 0 },
-      { fragment: 1, linkage: 'DYNAMIC', packageCount: 1, pkg: 1 },
-      { fragment: 2, linkage: 'DYNAMIC', packageCount: 0, pkg: 2 },
-    ],
-    packageCount: 3,
-    packages: [
-      { name: 'root', namespace: 'com.example', type: 'Maven', version: '1.0' },
-      {
-        name: 'library',
-        namespace: 'com.example',
-        type: 'Maven',
-        version: '2.0',
-      },
-      { name: 'leaf', namespace: 'com.example', type: 'Maven', version: '3.0' },
-    ],
-    purls: [
-      null as unknown as string,
-      'pkg:maven/com.example/library@2.0',
-      'pkg:maven/com.example/leaf@3.0',
-    ],
-    projectGroups: [],
-  };
-
-  describe('dependency graph helpers', () => {
-    it('formats dependency graph package labels from identifiers', () => {
-      expect(formatDependencyGraphPackageLabel(graph, 0, 'ORT_ID')).toBe(
-        'Maven:com.example:root:1.0'
-      );
-    });
-
-    it('returns an empty label for missing identifiers', () => {
-      const graphWithMissingPackage: DependencyGraph = {
-        ...graph,
-        packages: [undefined as unknown as DependencyGraph['packages'][number]],
-      };
-
-      expect(
-        formatDependencyGraphPackageLabel(graphWithMissingPackage, 0, 'ORT_ID')
-      ).toBe('');
-    });
-
-    it('matches a subtree when a descendant matches the search term', () => {
-      const matcher = createNodeSubtreeMatcher(
-        graph,
-        buildAdjacencyMap(graph),
-        'leaf',
-        'ORT_ID'
-      );
-
-      expect(matcher(0)).toBe(true);
-      expect(matcher(1)).toBe(true);
-      expect(matcher(2)).toBe(true);
-    });
-
-    it('does not recurse forever on cyclic graphs', () => {
-      const cyclicGraph: DependencyGraph = {
-        ...graph,
-        edges: [
-          { from: 0, to: 1 },
-          { from: 1, to: 0 },
-        ],
-      };
-      const matcher = createNodeSubtreeMatcher(
-        cyclicGraph,
-        buildAdjacencyMap(cyclicGraph),
-        'library',
-        'ORT_ID'
-      );
-
-      expect(matcher(0)).toBe(true);
-      expect(matcher(1)).toBe(true);
-    });
-
-    it('formats dependency graph package labels from purls when preferred', () => {
-      expect(formatDependencyGraphPackageLabel(graph, 1, 'PURL')).toBe(
-        'pkg:maven/com.example/library@2.0'
-      );
-      expect(formatDependencyGraphPackageLabel(graph, 0, 'PURL')).toBe(
-        'Maven:com.example:root:1.0'
-      );
-    });
-  });
-}
